Move ReviewOperation to components folder
