+++ webapp/app/[locale]/tunnel/_components/reviewOperation/operation.tsx
@@ -1,9 +1,9 @@
-import { DrawerSection } from 'components/drawer'
+import { ReviewOperation } from 'components/reviewOperation'
+import { Amount } from 'components/reviewOperation/amount'
+import { type StepPropsWithoutPosition } from 'components/reviewOperation/step'
 import { ComponentProps, ReactNode } from 'react'
 
-import { Amount } from './amount'
 import { Header } from './header'
-import { Step, type StepPropsWithoutPosition } from './step'
 
 type Props = {
   callToAction?: ReactNode
@@ -27,20 +27,11 @@ export const Operation = ({
     <div className="mb-3">
       <Header onClose={onClose} subtitle={subtitle} title={title} />
     </div>
-    <DrawerSection>
-      <Amount token={token} value={amount} />
-      <div
-        className="mt-4 flex flex-col gap-y-8"
-        // use this to prevent layout shift for the gray container
-        // calculate the max height depending on the number of steps
-        // and add 20px for the "total" section
-        style={{ height: `${144 * steps.length + 20}px` }}
-      >
-        {steps.map((stepProps, index) => (
-          <Step key={index} position={index + 1} {...stepProps} />
-        ))}
-      </div>
-    </DrawerSection>
-    {!!callToAction && <div className="h-full py-3">{callToAction}</div>}
+    <ReviewOperation
+      amount={amount}
+      callToAction={callToAction}
+      steps={steps}
+      token={token}
+    />
   </>
 )
