--- webapp/app/[locale]/tunnel/_components/reviewOperation/verticalLine.tsx
@@ -1,6 +1,0 @@
-export const VerticalLine = () => (
-  <div className="ml-4 py-1">
-    <div className="mb-3 h-2 w-[2px] bg-slate-100"></div>
-    <div className="h-2 w-[2px] bg-slate-100"></div>
-  </div>
-)
