+++ webapp/components/reviewOperation/_icons/orangeCheckIcon.tsx
