+++ webapp/components/reviewOperation/step.tsx
@@ -6,9 +6,9 @@ import { ComponentProps } from 'react'
 import { RemoteChain } from 'types/chain'
 import { getFormattedValue } from 'utils/format'
 
+import { ClockIcon } from './_icons/clockIcon'
+import { FeesIcon } from './_icons/feesIcon'
 import { OneRowBox, TwoRowBox } from './box'
-import { ClockIcon } from './icons/clockIcon'
-import { FeesIcon } from './icons/feesIcon'
 import { PositionStatus } from './positionStatus'
 import { ProgressStatus } from './progressStatus'
 import { SeeOnExplorer } from './seeOnExplorer'
